Refactor to use cobra, add version command
